+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 New feature(s):
 
 - New endpoint `/benchmark_score_stats` for querying benchmark config and score statistics.
+- Add optional `countries` and `vendor_regions` query filters for the `/server/{vendor}/{server}/prices` and
+  `/server/{vendor}/{server}/similar_servers/{by}/{num}` endpoints.
+- Add `benchmark_score` and `benchmark_score_per_price` ordering options to
+  `/server/{vendor}/{server}/similar_servers/{by}/{num}` endpoint.
+- Add optional `currency`, `benchmark_id`, `benchmark_config`, `server_region` and `best_price_allocation` query
+  parameters for the `/server/{vendor}/{server}/similar_servers/{by}/{num}` endpoint.
 
 Fix(es):